flask_app/requirements.txt
certifi
elastic-transport
elasticsearch
Flask
requests
urllib3
google-generativeai
